Use MinHook for win32 hooks
Instead of IAT hooking, just hook functions directly. This greatly simplifies the proof of concept since no LoadLibrary/GetProcAddress hooking is strictly necessary. renderdoc.dll loads in everything we'd want to hook, anyway, so only installing hooks at init time should be fine.  There's an additional hack here to work around a certain program's anti-tamper check -- the program checks the first byte of D3D11CreateDevice{,AndSwapChain} for a hook (0xE9) and bails if it sees it; this workaround is to hook the inner function in d3d11, but is done with manual offsets targeting the 64-bit d3d11 on my Windows 11 machine.  Two potential ways around this -- 1: replace MinHook with a hooking method that uses push imm16 / push imm32 / ret for the detour code, as this might be enough to fool the program; 2: disassemble the D3D11 targets forward enough to find the lea rcx, [impl_fn] op and derive the impl address that way (and perhaps the equivalent on 32-bit but let's be honest 32-bit should be killed).